+++ libraries/apollo-websocket-network-transport-incubating/src/commonTest/kotlin/com/apollographql/apollo3/network/websocket/WebSocketEngineTest.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.network.ws.incubating
+package com.apollographql.apollo3.network.websocket
 
 import com.apollographql.apollo3.exception.ApolloException
 import com.apollographql.apollo3.mockserver.CloseFrame
