+++ libraries/apollo-websocket-network-transport-incubating/src/jvmMain/kotlin/com/apollographql/apollo3/network/websocket/WebSocketEngine.jvm.kt
@@ -1,7 +1,9 @@
-package com.apollographql.apollo3.network.ws.incubating
+package com.apollographql.apollo3.network.websocket
 
 import com.apollographql.apollo3.api.http.HttpHeader
 import com.apollographql.apollo3.exception.ApolloNetworkException
+import com.apollographql.apollo3.network.websocket.WebSocket
+import com.apollographql.apollo3.network.websocket.WebSocketListener
 import okhttp3.Headers
 import okhttp3.OkHttpClient
 import okhttp3.Request
