+++ libraries/apollo-websocket-network-transport-incubating/src/wasmJsMain/kotlin/com/apollographql/apollo3/network/websocket/WebSocketEngine.wasmJs.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.network.ws.incubating
+package com.apollographql.apollo3.network.websocket
 
 actual fun WebSocketEngine(): WebSocketEngine {
   TODO("Not yet implemented")
